+++ lib/quaderno-ruby/helpers/response.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Quaderno::Helpers
+  module Response
+    def response_body=(response)
+      @response_body = response
+    end
+
+    def response_body
+      @response_body
+    end
+  end
+end